New "transfer" dialog window in the gtk interface.
git-svn-id: https://yabause.svn.sourceforge.net/svnroot/yabause/trunk@715 d7be59e3-070e-4b9a-85a1-4646befae34d

+++ yabause/src/linux/yui.c
@@ -98,7 +98,7 @@ static int yuiGetInt( const char* name, int deflt ) {
 
   char* res = yuiGetString( name );
   if ( !res ) return deflt;
-  return atoi( res );
+  return atol( res );
 }
 
 static void yuiSetInt( const char* name, const int value ) {
